services/sec_manager/requirements.txt
flask==2.2.5
pyjwt==2.6.0